requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.3
    # via requests
greenlet==3.2.4
    # via sqlalchemy
idna==3.11
    # via requests
joblib==1.5.2
    # via
    #   -r requirements.in
    #   scikit-learn
numpy==2.3.3
    # via
    #   pandas
    #   scikit-learn
    #   scipy
pandas==2.3.3
    # via -r requirements.in
psycopg2-binary==2.9.11
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
requests==2.32.5
    # via -r requirements.in
scikit-learn==1.7.2
    # via -r requirements.in
scipy==1.16.2
    # via scikit-learn
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.44
    # via -r requirements.in
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via sqlalchemy
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
